requirements.txt
ultralytics
albumentations
opencv-python
PyYAML
tqdm
matplotlib
